feat: 真会话切到新界面，旧的那一副整个删掉——终端循环每帧盯那一趟、画 `shell::draw`，按键经 `input` 交给新会话；删掉 `press`、`draw` 十四个模块连同快照用例、旧焦点那一维与旧按键表、`Live` 旧卷身份（`Volume`/`Branch`，一卷那一份改按卷根认）；颜色与设计快照读法搬进 `shell`，视口那张表换成新的七处；三条失败路径用例改经输入入口；CONTEXT.md 删《会话》节首那一句、ADR 0019 已落地；记 Q961–Q968

+++ src/session/shell/completions.rs
@@ -47,7 +47,7 @@ pub(super) fn draw(canvas: &mut Canvas<'_>, session: &Session) {
     if let Some(bar) = viewport.scrollbar() {
         canvas.scrollbar(area, &bar);
     }
-    let from = usize::from(viewport.from());
+    let from = viewport.from();
     for (i, candidate) in line.candidates.iter().enumerate().skip(from).take(shown) {
         let current = i == line.at;
         let name = if candidate.directory {
